feat: Add window level FMT fields and convert Int to Float
* Add window level metrics to SampleInfo struct
* Keep scores in float64 instead of casting down to int
* Adds WDC, WTC and PRF window level format fields
* Change AFR, SFS types from Integer to Float
* Rename HQ to AQ and change AQ type from Integer to Float

+++ src/lancet/hts/phred_quality.h
@@ -7,7 +7,7 @@ namespace lancet::hts {
 
 static constexpr u8 MAX_PHRED_SCORE = 255;
 [[nodiscard]] auto PhredToErrorProb(u32 phred_score) -> f64;
-[[nodiscard]] auto ErrorProbToPhred(f64 prob) -> u8;
+[[nodiscard]] auto ErrorProbToPhred(f64 prob) -> f64;
 
 }  // namespace lancet::hts
